# A quick fix for Stream operation errors on non-current device