@mrshenli 2019-01-03
Hi @ezyang thanks for reviewing the code.

Could you please help to click the rerun button of [this failed test](https://circleci.com/gh/pytorch/pytorch/504206). It was aborted for don't know what reason. I don't have a write permission to do so. Thanks.