@gchanan 2019-01-03
doesn't this issue exist on a bunch of functions (not just query?).  Is there some more holistic approach we could take here?