@mrshenli 2019-01-03
@gchanan 

I am thinking about implementing something like `torch._C._cuda_queryStream`, and replacing `cudaStreamQuery` invocations with the new implementation (not sure). I initially wanted to change `cudaStreamQuery` directly, but it seems to be an API in cuda which we cannot modify (maybe we can talk to NVIDIA??).