@mrshenli 2019-01-03
`test_streams_multi_gpu_query (test_cuda.TestCuda)` keep failing (no error message on ) rocmdeb, will skip the test on that platform.